jarditou php

+++ delete_form.php
@@ -1,19 +1,22 @@
 <?php
 ini_set('display_errors', 1); 
 error_reporting(E_ALL);  
-include("db.php");
+include_once('db.php');
+
 $pro_id = (int)$_GET['pro_id'];
 
             $query = $db->prepare("SELECT pro_id FROM produits WHERE pro_id = :pro_id ORDER BY pro_libelle");
             $query->bindParam(":pro_id", $pro_id);
             $query->execute();
             $row = $query->fetch(PDO::FETCH_ASSOC);
-
-if(!empty($row['pro_id']))
+            $query->closeCursor();//on libére la mémoire
+if(!empty($row['pro_id'])&&!empty($_GET['del'])&&$_GET['del']==1&&!empty($_GET['confirm'])&&$_GET['confirm']==1)
 {
             $req = $db->prepare('DELETE FROM produits WHERE pro_id= :pro_id');
             $req->bindParam(':pro_id', $row['pro_id']);
             $success = $req->execute();
+            $req->closeCursor();//on libére la mémoire
+
 
     if($success)
     {
@@ -30,3 +33,4 @@ else
 {
     header("Location: index.php");
 }
+?>
